Allow relative path for output-path and working-directory-path
Motivation:
The package-collection-generate command allows relative path for
input-path, but when used for output-path and working-directory-path,
the following error occurs.

Modifications:
Check if a path is absolute and in case it isn't, assume it is a
relative path relative to the current directory.

Result:
Using relative path doesn't result in error.