Add summary and createdAt to package version
Requires https://github.com/apple/swift-package-manager/pull/3266

The `generate` tool will try to obtain message and timestamp for **annotated** tags only. In most use-cases, tag is created as part of a (GitHub) release and is "lightweight", which is the opposite of annotated and the information we get with `git` would not be suitable.